% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpdust.R
\name{fpdust}
\alias{fpdust}
\title{Select a sample from a finite population using FPDUST procedure.}
\usage{
fpdust(spMat, n, beta)
}
\arguments{
\item{spMat}{a matrix representing spatial contingency among population units,
with dimension \code{NxN} and whose values are 1's (adjacent units) or 0's (otherwisex)}

\item{n}{integer scalar representing the sample size}

\item{beta}{a real number <1 representing a modifier of the probability of
selection for units that are adjacent to already selected units}
}
\value{
a logical vector indicating which units have been selected
        (TRUE if selected, FALSE otherwise)
}
\description{
Select a sample of size n from a finite population given a spatial
contingency matrix, the sample size \code{n}, and a parameter \code{beta}.
}
