% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialContMat.R
\name{spatialContingencyMatrix}
\alias{spatialContingencyMatrix}
\title{Create a spatial contingency matrix from a list of contingencies}
\usage{
spatialContingencyMatrix(clist)
}
\arguments{
\item{clist}{list of length N in which each element is a vector of id's continguous units}
}
\description{
Create a matrix of local contingency matrix of size \code{NxN} starting from
a list.
}
\details{
The input argument \code{clist} should be a list of lenght N. Each element of
this list must be a numeric vector, including the indices of the units that
are contiguous to the unit corresponding to the list element.
Function \code{spatialContingencyMatrix} returns a matrix whose values are 1's (contiguous units)
and 0's (non-contiguous units).
}
\examples{
# List of contiguous units ---
cl <- list( c(2, 3), c(1,4), c(1,6), c(2,6), NULL, c(3,4))

# Generate spatial contingency matrix ---
cm <- spatialContingencyMatrix(cl)


}
