% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialContMat.R
\name{spatialContMat}
\alias{spatialContMat}
\title{Create a spatial contingency matrix from a list of contingencies}
\usage{
spatialContMat(clist)
}
\arguments{
\item{clist}{list of length N in which each element is a vector of id's continguous units}
}
\description{
Given a \code{list} of contingencies for each unit in a population of size \code{N},
\code{scmat} creates a matrix of local contingency matrix of size \code{NxN}
of 1's (if units i and j are contiguous) and 0's (otherwise).
}
