% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpdust.R
\name{fpdust}
\alias{fpdust}
\title{Select a sample from a finite population using FPDUST procedure.}
\usage{
fpdust(spMat, n, beta)
}
\arguments{
\item{spMat}{spatial contingency matrix for population units, it should be a \code{NxN}
matrix of 1's (adjacent units) or 0's (otherwisex)}

\item{n}{integer representing sample size}

\item{beta}{a real number <1 representing a modifier of the probability of
selection for units that are adjacent to already selected units}
}
\value{
a vector indicating which units have been selected (1=selected, 0 otherwise)
}
\description{
\code{fpdust} selects a sample of size n from a finite population given a spatial
contingency matrix and a parameter \code{beta}.
}
