% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpdustPPS.R
\name{fpdustPPS}
\alias{fpdustPPS}
\title{Select a sample from a finite population using  a modified FPDUST procedure.}
\usage{
fpdustPPS(spMat, X, n, beta)
}
\arguments{
\item{spMat}{a matrix representing spatial contingency among population units,
with dimension \code{NxN} and whose values are 1's (adjacent units) or 0's (otherwisex)}

\item{X}{numeric vector with values of the size variable}

\item{n}{integer scalar representing sample size}

\item{beta}{a real number <1 representing a modifier of the probability of
selection for units that are adjacent to already selected units}
}
\value{
a logical vector indicating which units have been selected
        (TRUE if selected, FALSE otherwise)
}
\description{
\code{fpdustPPS} selects a sample of size n from a finite population given a spatial
contingency matrix, a parameter \code{beta}, and a size variable vector.
}
